Put Wallet types into separate module
Also introduce Hydra.Chain.Blockfrost.Wallet
